backend/requirements.txt
asgiref==3.5.2
certifi==2022.9.14
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.5.0
pip==22.2.2
psycopg2-binary==2.9.6
pytz==2022.4
setuptools==63.4.1
sqlparse==0.4.1
tzdata==2021.1
wheel==0.37.1
wincertstore==0.2
